rename main script and clean up format with black
