+++ eda.py
@@ -69,5 +69,5 @@ def eda(data: pd.DataFrame) -> None:
 
 
 if __name__ == "__main__":
-    data: pd.DataFrame = get_data(csv='data/sundae.csv')  # reminder: update file string
+    data: pd.DataFrame = get_data(csv="data/sundae.csv")  # reminder: update file string
     eda(data)
